Revert "Feat: 나에게 편지 보내기 예약 발송 기능 구현 (#67)"
