+++ .env.example
@@ -12,14 +12,6 @@ PASSPORT_NAVER_CLIENT_SECREAT={네이버 클라이언트 SECREAT}
 
 JWT_ACCESS_SECREAT={서명키}
 JWT_REFRESH_SECREAT={서명키}
-
-JWT_ACCESS_EXPIRED_TIME={시간}
-JWT_REFRESH_EXPIRED_TIME={시간}
-
-FRONTEND_CALLBACK_URL=http://localhost:3000
-NODE_ENV=develop
-
-WEEKLY_REPORT_FAIL_FAST=true
 #-----------------------DB 관련 환경변수 --------------------------
 DATABASE_URL={URL}
 REDIS_HOST="localhost"
@@ -40,29 +32,6 @@ MAILER_USER={메일서버 아이디}
 MAILER_PASS={메일서버 비밀번호}
 #---------------------------------------------------------------
 
-#---------------------------------------------------------------
-OCI_NAMESPACE={오브젝트 스토리지 네임스페이스}
-OCI_BUCKET_NAME={오브젝트 스토리지 버킷 이름}
-OCI_REGION={지역}
-OCI_COMPARTMENT_OCID={오브젝트 스토리지 컴파트먼트 OCID}
-#---------------------------------------------------------------
-
-#---------------------------------------------------------------
-OCI_TENANCY_OCID={테넌시 OCID}
-OCI_USER_OCID={유저 OCID}
-OCI_FINGERPRINT={핑거프린트}
-OCI_PRIVATE_KEY_PATH={프라이빗키}
-#OCI_PRIVATE_KEY_PASSPHRASE=
-#---------------------------------------------------------------
-
-#----------------------배포 CORS 환경변수 ----------------------------
-ALLOWED_ORIGINS={http://localhost:3000}
-#---------------------------------------------------------------
-
-#----------------------웹 push VAPID키-------------------------------
-VAPID_PUBLIC={VAPID 퍼블릭키}
-VAPID_PRIVATE={VAPID 프라이빗키}
-#---------------------------------------------------------------
 #------------------------GROQ---------------------------------
 GROQ_API_KEY={groq-api-key}
 GROQ_LETTER_MODEL=llama-3.3-70b-versatile
