+++ package.json
@@ -49,7 +49,6 @@
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.1",
     "uuid": "^13.0.0",
-    "web-push": "^3.6.7",
     "zod": "^4.3.5"
   },
   "devDependencies": {
