+++ src/Auths/strategies/naver.strategy.js
@@ -12,10 +12,9 @@ export const naverStrategy = new NaverStrategy(
 
     async (accessToken, refreshToken, profile, cb) => {
         try{
-            const {userId, tokens} = await verifySocialAccount(createSocialUserDTO(profile));
+            const {tokens} = await verifySocialAccount(createSocialUserDTO(profile));
 
             return cb(null, {
-                id: userId,
                 jwtAccessToken: tokens.jwtAccessToken,
                 jwtRefreshToken: tokens.jwtRefreshToken
             });
