--- src/configs/push.config.js
@@ -1,9 +1,0 @@
-import webpush from 'web-push';
-
-export const configurePush = () => {
-  webpush.setVapidDetails(
-    'mailto:soksakletter@gmail.com',
-    process.env.VAPID_PUBLIC_KEY,
-    process.env.VAPID_PRIVATE_KEY
-  );
-};
