+++ src/controllers/dashboard.controller.js
@@ -1,6 +1,5 @@
 import { getUserLetterStats } from "../services/letter.service.js";
 import { getTodayQuestion } from "../services/question.service.js";
-import { countUserSession } from "../services/session.service.js";
 
 export const HandleGetHomeDashboard = async (req, res, next) => {
     const userId = req.user.id;
@@ -8,12 +7,10 @@ export const HandleGetHomeDashboard = async (req, res, next) => {
     try{
         const todayQuestion = await getTodayQuestion();
         const letterStats = await getUserLetterStats(userId);
-        const sessionCount = await countUserSession(userId);
 
         res.status(200).success({
             todayQuestion, 
-            letterStats,
-            sessionCount
+            letterStats
         })
     } catch(err) {
         next(err);
