--- src/jobs/crons/letter.cron.js
@@ -1,16 +1,0 @@
-import cron from "node-cron";
-import { sendScheduledLetters } from "../bootstraps/letter.bootstrap.js";
-
-export const sendScheduledLettersCron = () => {
-    cron.schedule(
-        '0 0 * * *', 
-        async () => {
-            console.log("편지 전송을 시작합니다.");
-            sendScheduledLetters();
-        }, 
-        {
-            scheduled: false,
-            timezone: "Asia/Seoul"
-        }
-    );
-}
