+++ src/jobs/weeklyReport.bootstrap.js
@@ -1,14 +1,14 @@
 // jobs/weeklyReport.bootstrap.js
-import { prisma } from "../../configs/db.config.js";
-import { getISOWeek, getISOYear } from "../../utils/date.util.js";
-import { createWeeklyReport } from "../../services/weeklyReport.service.js";
+import { prisma } from "../configs/db.config.js";
+import { getCurrentISOYear, getCurrentISOWeek } from "../jobs/date.js";
+import { createWeeklyReport } from "../services/weeklyReport.service.js";
 
 // 동시에 몇 명까지 처리할지(너무 크게 잡으면 DB 터짐)
 const CONCURRENCY = Number(15);
 
 export const bootstrapWeeklyReports = async () => {
-  const year = getISOYear();
-  const week = getISOWeek();
+  const year = getCurrentISOYear();
+  const week = getCurrentISOWeek();
 
   const users = await prisma.user.findMany({
     where: { isDeleted: false },
