+++ src/repositories/question.repository.js
@@ -31,6 +31,4 @@ export const findQuestionByQuestionId = async(id) => {
             id: true
         }
     })
-
-    return result;
 }
