+++ src/services/session.service.js
@@ -9,7 +9,6 @@ import {
 import {
   SessionCountOverError,
   SessionInternalError,
-  SessionNotFoundError,
   SessionParticipantNotFoundError,
   UnExpectTagError,
   TemperatureRangeError,
@@ -45,12 +44,23 @@ export function validateTemperatureScore(temperatureScore) {
   return v >= 0 && v <= 100;
 }
 
-export const createMatchingSession = async (userId, targetUserId, questionId, tx) => {
-  try {
-    const result = await acceptSessionRequestTx(userId, targetUserId, questionId, tx);
-    if (result == null) throw new SessionInternalError(undefined, undefined, { userId, questionId });
+export const createMatchingSession = async (
+  userId,
+  questionId
+) => {
+  const count = await countMatchingSessionByUserId(userId);
+  console.log("count" + count);
+  if(count >= 10) throw new SessionCountOverError(undefined, undefined, { count });
+  const question = await findQuestionByQuestionId(questionId);
+  console.log("question" + question.id);
+  if(question == null) throw new QuestionNotFoundError(undefined, undefined, { questionId });
 
-    return result;
+  try {
+    const result = await acceptSessionRequestTx(userId, questionId);
+    if (result == null) throw new SessionInternalError(undefined, undefined, { userId, questionId });
+    return {
+      data: result,
+    };
   } catch (error) {
     if (error instanceof BadRequestError || error instanceof NotFoundError || error instanceof InternalServerError) {
       throw error;
@@ -135,9 +145,3 @@ export const createSessionReview = async (
     throw new SessionInternalError();
   }
 };
-
-export const countUserSession = (userId) => {
-  const count = countMatchingSessionByUserId(userId);
-
-  return count;
-}
